+++ packages/core/index.ts
@@ -6,5 +6,8 @@ export * from "./storage.js";
 export * from "./util.js";
 export * from "./safe-global.js";
 
+
+
 export * from "./timeline/index.js";
 export * from "./behaviors/index.js";
+export * from "./extendable.js";