+++ packages/core/timeline/timeline.ts
@@ -1,12 +1,8 @@
-import { broadcast } from "./moments.js";
-import { objectDefinedNotNull, isArray, isFunc } from "../util.js";
+import { broadcast, init } from "./moments.js";
+import { addObserver } from "./utils.js";
+import { objectDefinedNotNull, isArray } from "../util.js";
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const cloneDeep = require("lodash.clonedeep");
-
-// TODO:: work on these typings some more for improvements
-// TODO:: do we want to move to .env files, seems to be a sorta "norm" folks are using?
-
-export type ObserverAddBehavior = "add" | "replace" | "prepend";
 
 /**
  * Represents an observer that does not affect the timeline
@@ -40,8 +36,9 @@ type DistributeOn<T extends Moments, R extends Moments = T> =
     { [Prop in string & keyof T]: {
         (handler: Parameters<T[Prop]>[0][number]): Timeline<R>;
         toArray(): Parameters<T[Prop]>[0][number][];
-        replace(handler: ValidObserver): Timeline<R>;
-        prepend(handler: ValidObserver): Timeline<R>;
+        replace(handler: Parameters<T[Prop]>[0][number]): Timeline<R>;
+        prepend(handler: Parameters<T[Prop]>[0][number]): Timeline<R>;
+        clear(): boolean;
     }
     };
 
@@ -54,13 +51,14 @@ type DistributeEmit<T extends Moments> =
 /**
  * A type used to represent the proxied Timeline.clear property
  */
-type DistributeClear<T extends Moments> =
-    { [Prop in string & keyof T]: () => boolean };
+// type DistributeClear<T extends Moments> =
+//     { [Prop in string & keyof T]: () => boolean };
 
 /**
  * Virtual events that are present on all Timelines
  */
 type DefaultTimelineEvents<T extends Moments> = {
+    init: (observers: ((this: Timeline<T>) => Timeline<T>)[], ...args: any[]) => void;
     log: (observers: ((this: Timeline<T>, message: string, level: number) => void)[], ...args: any[]) => void;
     error: (observers: ((this: Timeline<T>, err: string | Error) => void)[], ...args: any[]) => void;
 };
@@ -78,7 +76,7 @@ type EmitProxyType<T extends Moments> = DistributeEmit<T> & DistributeEmit<Defau
 /**
  * The type combining the defined moments and DefaultTimelineEvents
  */
-type ClearProxyType<T extends Moments> = DistributeClear<T> & DistributeClear<DefaultTimelineEvents<T>>;
+// type ClearProxyType<T extends Moments> = DistributeClear<T> & DistributeClear<DefaultTimelineEvents<T>>;
 
 /**
  * Timeline represents a set of operations executed in order of definition,
@@ -90,10 +88,8 @@ export abstract class Timeline<T extends Moments> {
     private _parentObservers: ObserverCollection;
     private _onProxy: typeof Proxy | null = null;
     private _emitProxy: typeof Proxy | null = null;
-    private _clearProxy: typeof Proxy | null = null;
-    private _asyncOverride = false;
 
-    constructor(protected readonly moments: T, protected observers?: ObserverCollection) {
+    constructor(protected readonly moments: T, protected observers?: ObserverCollection, protected state: Record<symbol, any> = {}) {
 
         // TODO:: this work isn't correct
         if (objectDefinedNotNull(this.observers)) {
@@ -102,14 +98,6 @@ export abstract class Timeline<T extends Moments> {
             this._inheritingObservers = false;
             this.observers = {};
         }
-    }
-
-    public get AsyncOverride(): boolean {
-        return this._asyncOverride;
-    }
-
-    public set AsyncOverride(value: boolean) {
-        this._asyncOverride = value;
     }
 
     /**
@@ -121,7 +109,7 @@ export abstract class Timeline<T extends Moments> {
             this._onProxy = new Proxy(this, {
                 get: (target: any, p: string) => Object.assign((handler: ValidObserver) => {
 
-                    // // TODO:: we might need better logic here depending on how objects are constructed
+                    // // TODO:: we need better logic here depending on how objects are constructed
                     if (this._inheritingObservers) {
                         // ONLY clone the observers the first time this instance of timeline sets an observer
                         // this should work all up and down the tree.
@@ -132,6 +120,7 @@ export abstract class Timeline<T extends Moments> {
 
                     addObserver(target.observers, p, handler, "add");
                     return target;
+
                 }, {
                     toArray: (): ValidObserver[] => {
                         return Reflect.has(target.observers, p) ? cloneDeep(Reflect.get(target.observers, p)) : [];
@@ -146,34 +135,21 @@ export abstract class Timeline<T extends Moments> {
                         return target;
                         // Reflect.set(target, `__once${p}`, handler);
                     },
+                    clear: (): boolean => {
+
+                        if (Reflect.has(target.observers, p)) {
+                            // we trust outselves that this will be an array
+                            (<ObserverCollection>target.observers)[p].length = 0;
+                            return true;
+                        }
+
+                        return false;
+                    },
                 }),
             });
         }
 
         return <any>this._onProxy;
-    }
-
-    /**
-     * Property allowing access to subscribe observers to all the moments within this timline
-     */
-    public get clear(): ClearProxyType<T> {
-
-        if (this._clearProxy === null) {
-            this._clearProxy = new Proxy(this, {
-                get: (target: any, p: string) => () => {
-
-                    if (Reflect.has(target.observers, p)) {
-                        // we trust outselves that this will be an array
-                        (<ObserverCollection>target.observers)[p].length = 0;
-                        return true;
-                    }
-
-                    return false;
-                },
-            });
-        }
-
-        return <any>this._clearProxy;
     }
 
     /**
@@ -204,14 +180,6 @@ export abstract class Timeline<T extends Moments> {
             this._emitProxy = new Proxy(this, {
                 get: (target: any, p: string) => (...args: any[]) => {
 
-                    // check for once
-                    // TODO::
-                    // if (Reflect.has(target, `__once${p}`)) {
-                    //     const onceHandler = target[`__once${p}`];
-                    //     delete target[`__once${p}`];
-                    //     console.log("here");
-                    // }
-
                     // handle the case there are no observers registered to the target
                     const observers = Reflect.has(target.observers, p) ? Reflect.get(target.observers, p) : [];
 
@@ -224,7 +192,7 @@ export abstract class Timeline<T extends Moments> {
                     try {
 
                         // default to broadcasting any events without specific impl (will apply to defaults)
-                        const moment = Reflect.has(target.moments, p) ? Reflect.get(target.moments, p) : broadcast();
+                        const moment = Reflect.has(target.moments, p) ? Reflect.get(target.moments, p) : p === "init" ? init() : broadcast();
 
                         return Reflect.apply(moment, target, [observers, ...args]);
 
@@ -244,49 +212,30 @@ export abstract class Timeline<T extends Moments> {
         return <any>this._emitProxy;
     }
 
-    protected abstract execute(init?: RequestInit): Promise<any>;
-}
+    /**
+     * Starts a timeline
+     *
+     * @description This method first emits "init" to allow for any needed initial conditions then calls execute with any supplied init
+     *
+     * @param init A value passed into the execute logic from the initiator of the timeline
+     * @returns The result of this.execute
+     */
+    protected start(init?: any): Promise<any> {
 
-/**
- * Adds an observer to a given target
- *
- * @param target The object to which events are registered
- * @param moment The name of the moment to which the observer is registered
- * @param prepend If true the observer is prepended to the collection (default: false)
- *
- */
-function addObserver(target: Record<string, any>, moment: string, observer: ValidObserver, addBehavior: ObserverAddBehavior = "add"): any[] {
+        // TODO:: should we somehow create a copy of "this" so that any modifications (extensions, whatever)
+        // are left only to each execution of start vs. running it twice init is called twice and things could be double extended etc.
 
-    if (!isFunc(observer)) {
-        throw Error("Observers must be functions.");
+        // initialize our timeline
+        this.emit.init();
+
+        // execute the timeline
+        return this.execute(init);
     }
 
-    if (!Reflect.has(target, moment)) {
-
-        // if we don't have a registration for this moment, then we just add a new prop
-        Reflect.defineProperty(target, moment, {
-            value: [observer],
-            configurable: true,
-            enumerable: true,
-            writable: true,
-        });
-
-    } else {
-
-        // if we have an existing property then we follow the specified behavior
-        switch (addBehavior) {
-            case "add":
-                target[moment].push(observer);
-                break;
-            case "prepend":
-                target[moment].unshift(observer);
-                break;
-            case "replace":
-                target[moment].length = 0;
-                target[moment].push(observer);
-                break;
-        }
-    }
-
-    return target[moment];
+    /**
+     * Method orchestrating the emit calls for the moments defined in inheriting classes
+     *
+     * @param init A value passed into start from the initiator of the timeline
+     */
+    protected abstract execute(init?: any): Promise<any>;
 }