+++ packages/queryable/behaviors/browser-fetch.ts
@@ -5,13 +5,13 @@ export function BrowserFetch(): (instance: Queryable2) => Queryable2 {
 
     return (instance: Queryable2) => {
 
-        instance.on.send(function (this: Queryable2, url: URL, init: RequestInit): Promise<any> {
+        instance.on.send.replace(function (this: Queryable2, url: URL, init: RequestInit): Promise<any> {
 
             this.emit.log(`Fetch: ${init.method} ${url.toString()}`, LogLevel.Verbose);
 
             return fetch(url.toString(), init);
 
-        }, "replace");
+        });
 
         return instance;
     };